requirements.txt
apache-airflow==1.10.10
selenium==3.141.0
requests==2.22.0